Refactor GeoTrellis Polygonal Summary API
This API is based on a proof of concept found in
contrib.polygonal. We provide the following
implementationsfor Raster[Tile] and
Raster[MultibandTile]:
- Max
- Min
- Sum
- Mean
- StreamingHistogram

This is a change from the old API, which implemented the polygonal
summary handlers for Tile and MultibandTile. This was changed
because in practice, we are always performing polygonal summaries
over the Raster abstraction.

The helper methods in raster.summary.polygonal.[Single|Multi]bandTilePolygonalSummaryMethods
were removed, as the equivalent operation is now a one line call:
```
tile.polygonalMax(extent, polygon)
// is now...
raster.polygonalSummary(polygon, MaxVisitor)
```

The implementations of polygonal summary for Int were removed. This
includes the aggregate operations as well as for FastMapHistogram.
In practice, you always want to perform polygonal summaries on Double
because (ADD REASON).

Signed-off-by: CloudNiner <andy@finkmiller.com>