WIP: Move geotrellis-contrib raster PolygonalSummary implementation to core
We chose StreamingHistogram over MutableHistogram[Double] because
we force very little additional specificity (StreamingHistogram
is effectively a MutableHistogram[Double]) but we significantly
reduce the cognitive overhead of a user understanding the types
that our default implementation will work for. Same argument
for FastMapHistogram.

Add monoids for Streaming and FastMap histograms to ease
CellVisitor implementation for these types.

The CellVisitor param on the polygonalSummary() methods is explicit.
Every interesting part of a polygonal summary implementation happens
in the CellVisitor, so it makes sense to be clear about where that's
coming from.

CellVisitor is contravariant on R. This clarifies type signatures
for different CellVisitor implementations based on Tile and
MultibandTile by preventing us from having to provide the Tile type
as a type parameter at the call site.

Moves empty definition onto CellVisitor. The definition
of empty is within the context of the visitor state, so
it makes sense to define it where it's used.